Separate validator and parser files. Use both new classes to validate and parse in the main method
